runtime/native: stub jvm.h methods
These are the methods that the runtime provides to other native libraries. They're used by the OpenJDK native libraries, and seem to be loaded eagerly, so a bunch of stubs were necessary.

None of the stubs are hit yet, since the runtime still has a preference for the VM builtin methods.